+++ src/dmLib/designMarginsLib.py
@@ -1,13 +1,252 @@
 import numpy as np
-from typing import Dict, Any, AnyStr, List, Type
+from typing import Dict, Any, AnyStr, List, Type, Union
 import matplotlib.pyplot as plt
-from .uncertaintyLib import Distribution
+from .uncertaintyLib import Distribution,gaussianFunc
 from .utilities import check_folder
 
 """Design margins library for computing buffer and excess"""
-class MarginNode():
 
-    def __init__(self,label='',cutoff=0.9,buffer_limit=0):
+class FixedParam():
+    def __init__(self,value:Union[float,int,str],key:str,
+        description:str='',symbol:str=''):
+        """
+        Contains description of an input parameter to the MAN
+        is inherited by DesignParam
+
+        Parameters
+        ----------
+        value : Union[float,int,str]
+            the value of the input spec
+        key : str
+            unique identifier
+        description : str, optional
+            description string, by default ''
+        symbol : str, optional
+            shorthand symbol, by default ''
+        """
+
+        self.value          = value
+        self.description    = description
+        self.symbol         = symbol
+        self.key            = key
+        self.type           = type(value)
+
+    def __call__(self) -> Union[float,int,str]:
+        """
+        retrieve the value of the parameter
+
+        Returns
+        -------
+        Union[float,int,str]
+            The value of the parameter
+        """
+
+        return self.value # return the requested value
+
+class DesignParam(FixedParam):
+    def __init__(self,value:Union[float,int,str],key:str,
+        universe:Union[tuple,list],description:str='',symbol:str=''):
+        """
+        Contains description of an input parameter to the MAN
+        is inherited by DesignParam, and FixedParam
+
+        Parameters
+        ----------
+        value : Union[float,int,str]
+            the value of the input spec
+        key : str
+            unique identifier
+        universe : Union[tuple,list]
+            the possible values the design parameter can take, 
+            If tuple must be of length 2 (upper and lower bound)
+            type(value) must be float, or int
+        description : str, optional
+            description string, by default ''
+        symbol : str, optional
+            shorthand symbol, by default ''
+        """
+        super().__init__(value,key,description,symbol)
+        
+        if type(universe) == tuple:
+            assert len(universe) == 2
+            assert self.type in [float,int]
+        elif type(universe) == list:
+            assert len(universe) > 0
+
+        self.universe = universe
+
+class InputSpec():
+    def __init__(self,value: Union[float,int,Distribution,gaussianFunc],
+        key: str,description:str='',symbol:str='',cov_index=0):
+        """
+        Contains description of an input specification
+        could deterministic or stochastic
+        
+        Parameters
+        ----------
+        value : Union[float,int,Distribution,gaussianFunc]
+            the value of the input spec, 
+            if type is Distribution then a sample is drawn
+        key : str
+            unique identifier
+        description : str, optional
+            description string, by default ''
+        symbol : str, optional
+            shorthand symbol, by default ''
+        cov_index : int, optional
+            which random variable to draw from 
+            if multivariate distribution is provided, by default 0
+        """
+
+        self.value          = value
+        self.description    = description
+        self.symbol         = symbol
+        self.key            = key
+        self.cov_index      = cov_index
+        self.type           = type(value)
+        self._samples       = np.empty(0)
+
+        # Check if input spec if stochastic
+        if type(value) in [Distribution,gaussianFunc]:
+            self.stochastic = True
+        elif type(value) in [float,int]:
+            self.stochastic = False
+
+        # Check if input spec is co-dependant on another
+        if type(value) == gaussianFunc:
+            self.ndim = value.ndim
+        else:
+            self.ndim = 1
+
+        assert self.cov_index <= self.ndim
+
+    @property
+    def samples(self) -> np.ndarray:
+        """
+        sample vector getter
+
+        Returns
+        -------
+        np.ndarray
+            vector of sample observations
+        """
+        return self._samples # return 
+
+    @samples.setter
+    def samples(self,s:Union[float,np.ndarray]):
+        """
+        Appends value observation s to sample vector
+
+        Parameters
+        ----------
+        s : Union[float,np.ndarray]
+            value to append to sample vector
+        """
+        self._samples = np.append(self._samples,s)
+
+    def reset(self):
+        """
+        Resets the stored samples
+        """
+        self._samples = np.empty(0)
+
+    def view(self,xlabel='',savefile=None):
+        """
+        view 1D or 2D plot of probability distribution of excess
+        """
+
+        self.fig, self.ax = plt.subplots(figsize=(8, 3))
+        self.ax.hist(self.samples, bins=100, density=True)
+        self.ax.set_xlabel(xlabel)
+        self.ax.set_ylabel('density')
+
+        if savefile is not None:
+            # Save figure to image
+            check_folder('images/%s' %(self.key))
+            self.fig.savefig('images/%s/%s.pdf' %(self.key,savefile), 
+                format='pdf', dpi=200, bbox_inches='tight')
+
+        plt.show()
+
+    def __call__(self,N=1):
+        """
+        draw random samples from value
+
+        Parameters
+        ----------
+        N : int, optional
+            Number of random samples to draw
+            default is one sample
+
+        Returns
+        -------
+        np.ndarray
+            A 1D array of size N, where 
+            N is the number of requested samples
+        """
+
+        if self.stochastic:
+            assert self.value.samples.shape[1] >= N
+            samples = self.value.samples[self.cov_index,-N:] # retrieve last N samples
+        else:
+            samples = self.value * np.ones(N)
+        
+        self.samples = samples # store the samples inside instance
+        return samples # return the requested number of samples
+
+class Behaviour():
+    def __init__(self,key=''):
+        """
+        This class stores the method for calculating its outputs
+            - Intermediate parameters
+            - Performance parameters
+            - Decided values
+            - target thresholds
+
+        Parameters
+        ----------
+        key : str, optional
+            string to tag instance with, default = ''
+        """
+        self.key            = key
+        self.intermediate   = None
+        self.performance    = None
+        self.decided_value  = None
+        self.threshold      = None
+
+    def reset(self):
+        """
+        Resets the stored variables
+        """
+        self.intermediate   = None
+        self.performance    = None
+        self.decided_value  = None
+        self.threshold      = None
+
+    def __call__(self, *args, **kwargs):
+        """
+        The function that will be used to calculate the outputs of the behaviour model
+            - Can be a deterministic model
+            - Can be a stochastic model (by calling a defined dmLib.Distribution instance)
+        This method must be redefined by the user for every instance
+
+        Example
+        -------
+        >>> # [in the plugin file]
+        >>> from dmLib import Behaviour
+        >>> class myBehaviour(Behaviour):
+        >>>     def behaviour(self,r,d):
+        >>>         # some specific model-dependent behaviour
+        >>>         self.intermediate   = d
+        >>>         self.performance    = r*2+1 / d
+        >>>         self.decided_value  = r**2
+        >>>         self.threshold      = r/d
+        """
+        # default code for the default behaviour
+        return
+
+class MarginNode():
+    def __init__(self,key:str='',cutoff:float=0.9,buffer_limit:int=0,type:str='must_exceed'):
         """
         Contains description and implementation 
         of a Margin Node object which is the building block
@@ -21,17 +260,22 @@ class MarginNode():
         buffer_limit : float, optional
             lower bound for beginning of buffer zone,
             default = 0.0
-        label : str, optional
+        key : str, optional
             string to tag instance with, default = ''
+        type : str, optional
+            possible values('must_exceed','must_not_exceed'), by default 'must_exceed'
         """
 
-        self.label = label
-        self.cutoff = 0.9
-        self.buffer_limit = 0.0
-        self._target = np.empty(0)
+        self.key            = key
+        self.cutoff         = cutoff
+        self.buffer_limit   = buffer_limit
+        self.type           = type
+        self._target        = np.empty(0)
         self._decided_value = np.empty(0)
-        self._excess = np.empty(0)
-        self._excess_dist = None
+        self._excess        = np.empty(0)
+        self._excess_dist   = None
+
+        assert self.type in ['must_exceed','must_not_exceed']
 
     @property
     def target(self):
@@ -200,8 +444,8 @@ class MarginNode():
 
         if savefile is not None:
             # Save figure to image
-            check_folder('images/%s' %(self.label))
-            self.fig.savefig('images/%s/%s.pdf' %(self.label,savefile), 
+            check_folder('images/%s' %(self.key))
+            self.fig.savefig('images/%s/%s.pdf' %(self.key,savefile), 
                 format='pdf', dpi=200, bbox_inches='tight')
 
         plt.show()
@@ -232,13 +476,13 @@ class MarginNode():
 
         if savefile is not None:
             # Save figure to image
-            check_folder('images/%s' %(self.label))
-            self.figC.savefig('images/%s/%s.pdf' %(self.label,savefile), 
+            check_folder('images/%s' %(self.key))
+            self.figC.savefig('images/%s/%s.pdf' %(self.key,savefile), 
                 format='pdf', dpi=200, bbox_inches='tight')
 
         plt.show()
 
-    def __call__(self,decided_value,target_threshold):
+    def __call__(self,decided_value:np.ndarray,target_threshold:np.ndarray):
         """
         Calculate excess given the target threshold and decided value
 
@@ -255,5 +499,71 @@ class MarginNode():
         self.decided_value = decided_value # add to list of decided values
         self.target = target_threshold # add to list of targets
 
-        e = decided_value - target_threshold
+        if self.type == 'must_exceed':
+            e = decided_value - target_threshold
+        elif self.type == 'must_not_exceed':
+            e = target_threshold - decided_value
+        else:
+            raise Exception('Wrong margin type (%s) specified. Possible values are "must_Exceed", "must_not_exceed".' %(str(self.type)))
         self.excess = e # add to list of excesses
+
+class MarginNetwork():
+    def __init__(self,design_params:List[DesignParam],input_specs:List[InputSpec],
+        fixed_params:List[FixedParam],behaviours:List[Behaviour],
+        margin_nodes:List[MarginNode],key:str=''):
+        """
+        The function that will be used to calculate a forward pass of the MAN
+        and associated metrics of the MVM
+            - The first metric is change absorption capability (CAC)
+            - The second metric is the impact on performance (IoP)
+        This class should be inherited and the forward method should be implemented by the user
+        to describe how the input params are related to the output params (DDs,TTs, and performances)
+
+        Parameters
+        ----------
+        design_params : List[DesignParam]
+            list of DesignParam instances 
+        input_specs : List[InputSpec]
+            list of InputSpec instances 
+        fixed_params : List[FixedParam]
+            list of FixedParam instances 
+        behaviours : List[Behaviour]
+            list of Behaviour instances
+        margin_nodes : List[MarginNode]
+            list of MarginNode instances
+        key : str, optional
+            string to tag instance with, default = ''
+        """
+        self.design_params  = design_params
+        self.input_specs    = input_specs
+        self.fixed_params   = fixed_params
+        self.behaviours     = behaviours
+        self.margin_nodes   = margin_nodes
+        self.key            = key
+
+    def forward(self, *args, **kwargs):
+        """
+        The function that will be used to calculate a forward pass of the MAN
+        (design_params,input_specs,fixed_params) -> (excess,performance)
+        This method must be redefined by the user for every instance
+
+        Example
+        -------
+        >>> # [in the plugin file]
+        >>> from dmLib import MarginNetwork
+        >>> class myMarginNetwork(MarginNetwork):
+        >>>     def forward(self):
+        >>>         # some specific model-dependent behaviour
+        >>>         d1 = self.design_params[0]
+        >>>         s1 = self.input_specs[0]
+        >>>         p1 = self.fixed_params[0]
+        >>>         b1 = self.behaviours[0]
+        >>>         b2 = self.behaviours[1] # dependant on b1
+        >>>         e1 = self.margin_nodes[0]
+        >>>         # Execution behaviour models
+        >>>         b1(d1.value,p1.value)
+        >>>         b2(d1.value,b1.intermediate)
+        >>>         e1(b3.decided_value,s3())
+        """
+        # default code for the default threshold
+        return